DixelKit/requirements.txt
pydicom
python-dateutil
pyyaml
python-magic
requests
splunk-sdk
aenum
beautifulsoup4